requirements.txt
aiohttp>=3.9.0,<4.0.0
cachetools>=5.3.0,<6.0.0
discord.py>=2.3.0,<3.0.0
